server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
# AI / CV
ultralytics==8.4.33
onnxruntime>=1.16.0
paddleocr==2.7.3
# Use contrib only — it's a superset of opencv-python (avoids module conflict)
opencv-contrib-python==4.6.0.66
numpy==1.26.4
pillow>=10.0.0
protobuf==3.20.2
shapely==2.1.2
scikit-image==0.26.0
pyclipper==1.4.0
lmdb==2.2.0
tqdm==4.67.3
rapidfuzz==3.14.3
cython==3.2.4
lxml==6.0.2
premailer==3.10.0
openpyxl==3.1.5
python-docx==1.2.0
beautifulsoup4==4.14.3
fire==0.7.1
pdf2docx==0.5.12
PyMuPDF>=1.24.0
pyyaml==6.0.3
requests==2.33.1
scipy==1.15.0
matplotlib>=3.8.0
pandas>=2.0.0
